was added negative tests

+++ app/routers/users.py
@@ -17,7 +17,6 @@ def get_user(user_id: int) -> User:
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
     user = users.get_user(user_id)
-
     if not user:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
     return user
@@ -46,5 +45,8 @@ def update_user(user_id: int, user: User) -> User:
 def delete_user(user_id: int):
     if user_id < 1:
         raise HTTPException(status_code=HTTPStatus.UNPROCESSABLE_ENTITY, detail="Invalid user id")
+    user = users.get_user(user_id)
+    if not user:
+        raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="User not found")
     users.delete_user(user_id)
     return {"message": "User deleted"}
